Reorganize project structure [WIP]
