+++ client/src/app/components/App/App.jsx
@@ -3,14 +3,13 @@ import { connect } from 'react-redux';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
 
-import Footer from 'modules/legacy/Footer';
-import Header from 'modules/legacy/Header';
-import MessageBox from 'modules/legacy/MessageBox';
-import ProductsContainer from 'modules/legacy/ProductsContainer';
-import Preloader from 'modules/legacy/Preloader';
+import Footer from 'app/components/Footer';
+import MessageBox from 'common/components/MessageBox';
+import ProductsContainer from 'modules/shopping-list/ProductsContainer';
+import Preloader from 'common/components/Preloader';
 import { StatusType, MessageType } from 'common/constants/enums';
 import { getItems, getFetchStatus } from 'modules/legacy/selectors';
-import InputBar from 'modules/legacy/InputBar';
+import InputBar from 'modules/shopping-list/input-bar';
 import { StatusPropType } from 'common/constants/propTypes';
 import { fetchItems } from 'modules/legacy/appActions';
 import Toolbar from '../Toolbar/Toolbar';
@@ -27,7 +26,6 @@ class App extends Component {
 
   render() {
     const { fetchStatus, items } = this.props;
-
     const reversedItem = [...items].reverse();
     const archiveList = reversedItem.filter(item => item.isOrdered);
     const shoppingList = reversedItem.filter(item => !item.isOrdered);
@@ -40,7 +38,6 @@ class App extends Component {
             overlay: fetchStatus === StatusType.PENDING
           })}
         >
-          <Header />
           {fetchStatus === StatusType.ERROR && (
             <MessageBox
               message="Fetching failed. Try to refresh the page."
