+++ client/src/app/components/Toolbar/Toolbar.jsx
@@ -7,6 +7,7 @@ import UsersIcon from 'assets/images/users-solid.svg';
 import ListIcon from 'assets/images/clipboard-list-solid.svg';
 import PlusIcon from 'assets/images/plus-solid.svg';
 import UserBar from 'modules/legacy/UserBar';
+import AppLogo from 'common/components/AppLogo';
 
 const Toolbar = () => (
   <div className="toolbar">
@@ -47,6 +48,7 @@ const Toolbar = () => (
         <img alt="Plus Icon" className="toolbar__icon-plus" src={PlusIcon} />
       </a>
     </div>
+    <AppLogo />
     <div className="toolbar__right">
       <UserBar />
     </div>
