+++ client/src/common/components/AppLogo.jsx
