+++ client/src/common/components/MessageBox.jsx
