+++ client/src/modules/legacy/styles/containers/_app.scss
@@ -15,8 +15,9 @@ body {
 }
 
 .app-wrapper {
+  margin-top: 32px;
   max-width: 560px;
-  margin: 0 auto;
+  margin: 32px auto 0;
   position: relative;
   height: 100%;
 }
