+++ client/src/modules/shopping-list/FilterBox.jsx
