+++ client/src/modules/shopping-list/ProductsListItem.jsx
@@ -3,7 +3,7 @@ import classNames from 'classnames';
 import PropTypes from 'prop-types';
 
 import { PLACEHOLDER_URL } from 'common/constants/variables';
-import VotingBox from './VotingBox';
+import VotingBox from 'modules/shopping-list/VotingBox';
 
 const ProductListItem = ({
   archived,
