+++ client/src/modules/shopping-list/VotingBox.jsx
