+++ client/src/modules/shopping-list/input-bar/InputBar.jsx
@@ -5,7 +5,7 @@ import PropTypes from 'prop-types';
 import { getNewItemStatus, getCurrentUser } from 'modules/legacy/selectors';
 import { StatusType, MessageType } from 'common/constants/enums';
 import { StatusPropType, UserPropType } from 'common/constants/propTypes';
-import MessageBox from '../MessageBox';
+import MessageBox from 'common/components/MessageBox';
 import { addItem } from './actions';
 
 class InputBar extends Component {
