+++ client/src/modules/shopping-list/input-bar/index.js
