+++ client/src/modules/shopping-list/product-list/index.js
